Updating the libsecret logic and code.
