+++ src/rc.c
@@ -40,10 +40,11 @@
 #include <sys/stat.h>
 #include <fcntl.h>
 #include "gettext.h"
-#include "defs.h"
 #include "decrypt.h"
-#include "options.h"
+#include "defs.h"
 #include "mem.h"
+#include "options.h"
+#include "prompt.h"
 #include "secret.h"
 
 #ifndef LINE_MAX
@@ -148,6 +149,12 @@ read_rc(const char *file)
 				} else {
 					options.netrc = 0;
 				}
+			} else if (strncmp("libsecret", vals[0], 9) == 0) {
+				if ((vals[1][0] == 'y') || (vals[1][0] == 'Y')) {
+					options.libsecret = 1;
+				} else {
+					options.libsecret = 0;
+				}
 			} else if (strncmp("password_file", vals[0], 13) == 0) {
 				len = strlen(vals[1]) +1;
 				pfile = xmalloc(len);
@@ -183,15 +190,23 @@ read_rc(const char *file)
 		options.username = getenv("USER");
 	}
 
+	if (options.pwprompt) {
+		prompt_password();
+	}
+
 #if HAVE_LIBSECRET
-	if (!options.pwprompt) {
-		lookup_password();
+	if (options.libsecret) {
+		if (!options.pwprompt) {
+			if (lookup_password() == 1) {
+				return(EXIT_FAILURE);
+			}
+		} else {
+			store_password();
+		}
 	}
 #endif
 
-	if (options.password == NULL &&
-	    !options.pwprompt &&
-	    pfile) {
+	if (options.password == NULL && pfile) {
 #if HAVE_GPGME == 1
 		if (pfile[0] == '~' && pfile[1] == '/') {
 			home = getenv("HOME");
@@ -247,7 +262,6 @@ read_rc(const char *file)
 #endif
 	}
 	if (options.password == NULL &&
-	    !options.pwprompt &&
 	    options.netrc == 1) {
 		home = getenv("HOME");
 		if (home == NULL) {
@@ -272,15 +286,14 @@ read_rc(const char *file)
 		}
 	}
 
-	if ((options.pwprompt = options.password == NULL ? 1 : 0)) {
-		prompt_password();
 #if HAVE_LIBSECRET
+	if (options.libsecret == 1) {
 		if (options.password && options.password[0] == '\0') {
 			clear_password();
 			options.password = NULL;
 		}
-#endif
 	}
+#endif
 
 	return(EXIT_SUCCESS);
 }
